refactor: refactor search content
