+++ src/app/(normal)/search/[...slug]/pages/SearchPostContent.tsx
@@ -1,0 +1,64 @@
+'use client';
+
+import { useSuspenseInfiniteQuery } from '@tanstack/react-query';
+import { compact } from 'lodash-es';
+
+import { ListInPage } from '@/components/ListInPage.js';
+import { SinglePost } from '@/components/Posts/SinglePost.js';
+import { Empty } from '@/components/Search/Empty.js';
+import { SearchSources } from '@/components/Search/SearchSources.js';
+import { ScrollListKey } from '@/constants/enum.js';
+import { narrowToSocialSource } from '@/helpers/narrowToSocialSource.js';
+import { createIndicator } from '@/helpers/pageable.js';
+import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
+import type { Post } from '@/providers/types/SocialMedia.js';
+import { useSearchStateStore } from '@/store/useSearchStore.js';
+
+const getSearchItemContent = (post: Post, index: number, listKey: string) => {
+    return <SinglePost key={post.postId} post={post} listKey={listKey} index={index} />;
+};
+
+export function SearchPostContent() {
+    const { searchKeyword, searchType, source } = useSearchStateStore();
+    const currentSocialSource = narrowToSocialSource(source);
+
+    const queryResult = useSuspenseInfiniteQuery({
+        queryKey: ['search', searchType, searchKeyword, source],
+        queryFn: async ({ pageParam }) => {
+            if (!searchKeyword) return;
+            const provider = resolveSocialMediaProvider(currentSocialSource);
+            const indicator = pageParam ? createIndicator(undefined, pageParam) : undefined;
+
+            return provider.searchPosts(searchKeyword.replace(/^#/, ''), indicator);
+        },
+        initialPageParam: '',
+        getNextPageParam: (lastPage) => {
+            if (lastPage?.data.length === 0) return;
+            return lastPage?.nextIndicator?.id;
+        },
+        select(data) {
+            return compact(data.pages.flatMap((x) => x?.data ?? []));
+        },
+    });
+
+    const listKey = `${ScrollListKey.Search}:${searchType}:${searchKeyword}:${source}`;
+
+    return (
+        <>
+            <SearchSources source={source} query={searchKeyword} searchType={searchType} />
+            <ListInPage
+                source={source}
+                key={listKey}
+                queryResult={queryResult}
+                VirtualListProps={{
+                    listKey,
+                    computeItemKey: (index, post) => `${post.postId}_${index}`,
+                    itemContent: (index, post) => getSearchItemContent(post, index, listKey),
+                }}
+                NoResultsFallbackProps={{
+                    message: <Empty keyword={searchKeyword} />,
+                }}
+            />
+        </>
+    );
+}
